Core/Instance: stop updating the instance resettimes based on creature respawns
Port From (https://github.com/TrinityCore/TrinityCore/commit/7dcc185a151745c8a12e23708bb5d457e5796032)

+++ Source/Game/Entities/Player/Player.DB.cs
@@ -1030,7 +1030,7 @@ namespace Game.Entities
                     byte difficulty = result.Read<byte>(3);
                     BindExtensionState extendState = (BindExtensionState)result.Read<byte>(4);
 
-                    long resetTime = result.Read<uint>(5);
+                    long resetTime = result.Read<long>(5);
                     // the resettime for normal instances is only saved when the InstanceSave is unloaded
                     // so the value read from the DB may be wrong here but only if the InstanceSave is loaded
                     // and in that case it is not used
